requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
idna==3.10
peewee==3.18.1
peewee-migrate==1.13.0
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.5.0
